revisionstore: add batch indexedlog insertion helper
Summary:
Add "append_batch" which is optimized to avoid lock contention by only acquiring the indexedlog write lock once for the entire batch. It also supports filtering out items that are already in the log, which can be a big time (and space) saver if the item is likely to already be in the log.

Note that the underlying indexedlog does not support batch insertions - append_batch probably won't help very much in single threaded scenarios.

Reviewed By: quark-zju, zzl0

Differential Revision: D78748380

fbshipit-source-id: d1e8a41aeec2d98aacf1c40192c44020d9e87aa3