walkdetector: optimize perf of listing all active walks
Summary:
Add a new bool WalkNode.descendant_might_have_walk which tracks whether any descendant node might have a walk. This allows a fast path when traversing the tree looking for active walks.

During very large walks, I noticed we spend a significant amount of time listing the walks because the tree of WalkNodes is so huge. The vast majority of nodes do not "lead to" a walk, so this optimization is extremely effective to minimize work.

Reviewed By: quark-zju

Differential Revision: D78748386

fbshipit-source-id: e2697201c0cb768088493f0bb3ecaf5c7aaedc1b