chg: make chg server namespaced by cgroup id
Summary:
Add cgroup id to the socket name so the client/server don't cross cgroup boundaries.

The cgroup id is the inode of the (unified) cgroup's /sys/fs/cgroup/<name> directory. This assumes we are using cgroupv2 where a process is only ever in a single "unified hierarchy" cgroup.

This fixes a class of problems where your sapling command (and anything it spawns) can be running in a "random" cgroup with different limits/lifecycles, depending on who first spawned the pfc server.

Reviewed By: jtbraun, quark-zju

Differential Revision: D79201365

fbshipit-source-id: d1d3ea2c01451f0ab8ce48a2798172bfe94fbe6f