don't check the eden version in eden doctor when used with `--fast` or `--dry-run` (when doctor is ran via automation)
Summary:
The comment and T232917962 + T231826134 outline the reason for this change. tl;dr is that the OutOfVersion check can be confusing when ran and reported by automation, because as it exists today, it can be a little misleading/aggressive if we've not had a release in a while.

T232917962 contains the pseudocode for an ideal fix, which is to make the threshold dynamic rather than hardcoded to 14 days, but attempting to skip this codepath in the meantime for automation is an easy QOL fix.

Differential Revision: D79299691

fbshipit-source-id: 32d3ed8d5a36ac437710c9f578f7bed3584edd2a