Resolve all comments action
Summary:
This diff adds the next smart action, which prompts Devmate to resolve all comments on the selected diff. This is different from the existing "resolve comments" prompt, which only resolves a single comment. For that reason, I created a new prompt which borrows heavily from the existing one.

Currently I have this set up to only work via VS Code. At the very least, it doesn't make sense to prompt the Devmate extension from the browser (unless we open a new tab / prompt the browser extension?). I think that we could potentially add support for other platforms that have the ability to talk to Devmate (e.g. Android Studio), but I'd need help setting that up as I'm not sure how those platforms work.

Like other smart actions, this feature is gated behind a GK, in this case the same one that the inline comments webview uses for its Devmate feature. We could consider creating a separate GK for this, but we also have an overarching GK for the entire menu which gives some additional control during rollout.

Once again, looking for feedback on this approach!

Differential Revision: D79196458

fbshipit-source-id: f028ef0ead99f922d874305ac3df79caa4f1d832