Auto-split action
Summary:
AI split already exists within ISL via the split modal, so we can easily add an action to invoke this split functionality. Right now I'm using a somewhat hacky method to invoke this:

1. Set a custom boolean atom when the smart action is clicked.
2. Pop open the split modal.
3. Read the custom atom, and if set click the "AI split" button via ref.
4. Reset the custom atom.

The major advantage of this method is that it minimizes duplicated logic. We could also take an alternative approach (suggestion from stbfb) where we run the split first in the background (we might need to show progress in this case?), and then pop open the modal with the split result already loaded. This, however, would require a more involved solution.

Looking for feedback on this approach.

Reviewed By: stbfb

Differential Revision: D78989374

fbshipit-source-id: c4ccfd91fbe1329ab6a0902a2adbab33a885ec50