Show button while tooltip is visible
Summary:
An issue with the smart actions button is that it disappears if the commit it's rendered on is no longer hovered on. This is particularly problematic when the tooltip showing the smart actions menu is open, since it's not clear where it's being sourced from.

To fix this, I added a property to specify a function to run when a tooltip first becomes visible (e.g. updating some local state), similar to `onDismiss`. We can then use that local state to conditionally apply a CSS class.

Reviewed By: quark-zju

Differential Revision: D79196456

fbshipit-source-id: fa7f5bcc93755d11801e92241078672fabfb8ac7