Basic smart actions menu
Summary:
This diff adds the basic component for the consolidated "smart actions" menu. Design-wise, this is pretty similar to the "bulk actions" menu in the topbar, so I'm mostly following along with the implementation of that menu.

To reduce visual clutter (and to align with where the button appears in the mock-ups), I've made it so the "lightbulb" button only appears on hover.

Right now the menu only supports a `Commit` as the reference object for showing relevant smart actions, since this is the most clearly-defined usecase. The design doc touches on other types of context (uncommited changes, merge conflict state, comparison view), which will need to be handled differently.

Reviewed By: stbfb

Differential Revision: D78989370

fbshipit-source-id: 709940247c459162fbc22fb637302ee83493719a