sql_ext: Store shard_name in Sqlite connections
Summary:
## This stack

This stack will introduce a `Connection` wrapper in mononoke's `sql_ext` crate, to store the name of the database the connection belongs to. This is similar to what was done for `Transaction` in D77998324.

**Why do we need this?**

Right now, we're logging metadata about **queries to ALL our XDBs** (e.g. commit cloud, blobstore), which we don't need nor want. We want to get the data about cpu load, number of queries, etc specifically for `xdb.mononoke_production`.

### Solution
This wrapper will be pretty much like the `Transaction` wrapper introduced in D77998324.

The entire Mononoke codebase should use it instead of the one from the `sql` crate.

FYI, we also have to introduce wrappers for `SqlConnections` and `SqlShardedConnections`.

## This diff
Similar to D79443185, this will update the remaining callsites that were passing `None` as the shard_name to the Connection and Transaction wrappers.

This will allow us to make the field required in both wrappers in the next diff (D79444409).

Reviewed By: RajivTS

Differential Revision: D79443508

fbshipit-source-id: 6377f345492e1edcdb8a5c2a1f259817a3c947bc