sql_ext: Upsample errors
Summary:
## This stack

This stack will introduce a `Connection` wrapper in mononoke's `sql_ext` crate, to store the name of the database the connection belongs to. This is similar to what was done for `Transaction` in D77998324.

**Why do we need this?**

Right now, we're logging metadata about **queries to ALL our XDBs** (e.g. commit cloud, blobstore), which we don't need nor want. We want to get the data about cpu load, number of queries, etc specifically for `xdb.mononoke_production`.

### Solution
This wrapper will be pretty much like the `Transaction` wrapper introduced in D77998324.

The entire Mononoke codebase should use it instead of the one from the `sql` crate.

FYI, we also have to introduce wrappers for `SqlConnections` and `SqlShardedConnections`.

## This diff
D79444775 will significantly improve retention, since we'll be downsampling data for DBs we don't care about **and generate a ton of data** (e.g. XDB blobstore).

But one thing I also wanted to do is to ensure that we get the highest amount of data from errors that we can.

So I created a separate JK to set the sample rate for errors, which is also gated by shard_name. I initially set it to 1 (i.e. no sampling), but it should be easy to change if we want to.

Reviewed By: RajivTS

Differential Revision: D79447496

fbshipit-source-id: ee24098b6db9e5be8151dd546da627b72dac4e38