sql_ext: Use shard_name as switch value to get telemetry sample rate
Summary:
## This stack

This stack will introduce a `Connection` wrapper in mononoke's `sql_ext` crate, to store the name of the database the connection belongs to. This is similar to what was done for `Transaction` in D77998324.

**Why do we need this?**

Right now, we're logging metadata about **queries to ALL our XDBs** (e.g. commit cloud, blobstore), which we don't need nor want. We want to get the data about cpu load, number of queries, etc specifically for `xdb.mononoke_production`.

### Solution
This wrapper will be pretty much like the `Transaction` wrapper introduced in D77998324.

The entire Mononoke codebase should use it instead of the one from the `sql` crate.

FYI, we also have to introduce wrappers for `SqlConnections` and `SqlShardedConnections`.

## This diff
We are currently using a ton of storage and still have very little retention (see [idata](https://www.internalfb.com/intern/data/info/tables/scuba/uber/mononoke_sql_telemetry/)) because we're logging telemetry **for all DBs, including the XDB blobstore**...

We currently don't need any telemetry for the blobstore and the other DBs (e.g. commit cloud), so let's use the DB name as a switch value and as soon as these changes are released, I'll update the JK to lower the sampling rate of `xdb.mononoke_production` and increase the default.

Reviewed By: RajivTS

Differential Revision: D79444775

fbshipit-source-id: 246960ed855b950ab7713e03a799c206e6b52241