Update all callsites that create Transactions to use new constructor
Summary:
## This stack

This stack will introduce a `Connection` wrapper in mononoke's `sql_ext` crate, to store the name of the database the connection belongs to. This is similar to what was done for `Transaction` in D77998324.

**Why do we need this?**

Right now, we're logging metadata about **queries to ALL our XDBs** (e.g. commit cloud, blobstore), which we don't need nor want. We want to get the data about cpu load, number of queries, etc specifically for `xdb.mononoke_production`.

### Solution
This wrapper will be pretty much like the `Transaction` wrapper introduced in D77998324.

The entire Mononoke codebase should use it instead of the one from the `sql` crate.

FYI, we also have to introduce wrappers for `SqlConnections` and `SqlShardedConnections`.

## This diff

Updates all the callsites that create the `Transaction` wrapper by
1. creating a SQL transaction and
2. manually creating a Transaction wrapper passing everything needed

to just use a new instance method in the `Connection` wrapper that creates the `Transaction` wrapper when it's given the `SqlQueryTelemetry`.

This diff should be a no-op, just updating callsites.

Reviewed By: markbt

Differential Revision: D79435164

fbshipit-source-id: 3a3f2ac70f8342ec931a546b5793abae30007f5b