Log db_name
Summary:
## This stack

This stack will introduce a `Connection` wrapper in mononoke's `sql_ext` crate, to store the name of the database the connection belongs to. This is similar to what was done for `Transaction` in D77998324.

**Why do we need this?**

Right now, we're logging metadata about **queries to ALL our XDBs** (e.g. commit cloud, blobstore), which we don't need nor want. We want to get the data about cpu load, number of queries, etc specifically for `xdb.mononoke_production`.

### Solution
This wrapper will be pretty much like the `Transaction` wrapper introduced in D77998324.

The entire Mononoke codebase should use it instead of the one from the `sql` crate.

FYI, we also have to introduce wrappers for `SqlConnections` and `SqlShardedConnections`.

## This diff
With the shard name available in both Connection and Transaction wrappers, let's log it.

Reviewed By: markbt

Differential Revision: D79434109

fbshipit-source-id: 9c3af6fc4c18ca0b00df8d5672bb43be4f409e18