Use oneshot to signal termination
Summary: Currently, we busy wait on an atomic bool to ensure that we get the signal for termination in the sharding framework. That is less than ideal because we wake up our task every few seconds, realise nothing needs to be done and then go back to sleep. A better approach would be to use a `oneshot` that can be awaited without a loop and once the oneshot fires, we know its time to exit. This is what this diff does

Differential Revision: D79439888

fbshipit-source-id: cecc6c77a8ac5c51e77260a20fa8b89e7cc73b20