Remove ignore_artifacts for test rules dependent on Thrift
Summary:
This makes content-based paths possible for Thrift-generated files in fbcode. There is a potential small loss in build-time performance here, but these rules seem: (a) only used in tests; and (b) non-critical even in those tests. The benefits of having content-based Thrift rules should far outweigh any potential loss here IMO.

Long version: without `ignore_artifacts`, the action would need to wait for the input files to be built first instead of relying on their pre-determined paths (not available with content-based paths anymore), but this is generally not a big issue as they will be built anyway before they can be used. The only potential loss is from a slight out-of-order execution advantage, but this effect is often negligible to begin with.

Reviewed By: IanChilds

Differential Revision: D79546055

fbshipit-source-id: 44be5fd4353844ce171fdbb52236acda8e18843a