lfs_server: log to scuba if client attempted to fetch from CAS
Summary:
Context: We want to log SLAPI requests that reached Mononoke even though CASC was enabled. This should be a small number of requests since most requested tree and blobs should be CAS.

Add "fetch_from_cas_attempted" field to multiplexed blobstore scuba logs if the operation was in the context of a SLAPI request where the client attempted to fetch from CAS first.

Reviewed By: clara-9

Differential Revision: D79358457

fbshipit-source-id: 6358b7bbd7194bc969f478a693e8f3c117aa5012