add fetch_from_cas_attempted boolean to FetchContext
Summary:
Context: We want to log SLAPI requests that reached Mononoke even though CASC was enabled. This should be a small number of requests since most requested tree and blobs should be CAS.

Add a `fetch_from_cas_attempted` boolean to `FetchContext` and set it to true if fetching from cas was attempted.

Reviewed By: lmvasquezg

Differential Revision: D79358437

fbshipit-source-id: 07ba78689af28f23eb7cebb781b63b759cafd3be