server: log to scuba if client attempted to fetch from CAS
Summary:
Context: We want to log SLAPI requests that reached Mononoke even though CASC was enabled. This should be a small number of requests since most requested tree and blobs should be CAS.

Log in Mononoke server if the client attempted to fetch from CAS before making a SLAPI request.

Reviewed By: lmvasquezg

Differential Revision: D79358434

fbshipit-source-id: 5266b58d1d275809e2790c70ed4b347df848cce6