multiplexedblob: log to scuba if client attempted to fetch from CAS
Summary:
Context: We want to log SLAPI requests that reached Mononoke even though CASC was enabled. This should be a small number of requests since most requested tree and blobs should be CAS.

Add "fetch_from_cas_attempted" field to multiplexed blobstore scuba logs if the operation was in the context of a SLAPI request where the client attempted to fetch from CAS first.

Reviewed By: clara-9

Differential Revision: D79358459

fbshipit-source-id: c8e78a7f2efa094e4011c8ddd4f6e21c03c38d1d