Create MutableRepoBlobstore facet
Summary:
# This stack
I'm creating a separate blobstore for mutable data. See https://docs.google.com/document/d/1-hFZ9JFEMgfczCVM_UL2IPoXBpacdE4EC-7UwukpXvU/edit?tab=t.0 for the expected impact and detials on execution.

# This diff
In this diff we create the facet, but it's not yet used anywhere. The facet implementation is almost idential to `RepoBlobstore` with the exception of redaction config, which is not necessary: redaction applies to immutable data, data that is part of the commit graph.

Reviewed By: YousefSalama

Differential Revision: D78811362

fbshipit-source-id: a1002b99ad59e5a12e478b3fb704e5db6e6d04fd