Use MutableRepoBlobstore for Microwave
Summary:
# This stack
I'm creating a separate blobstore for mutable data. See https://docs.google.com/document/d/1-hFZ9JFEMgfczCVM_UL2IPoXBpacdE4EC-7UwukpXvU/edit?tab=t.0 for the expected impact and detials on execution.

# This diff
In this diff we use the facet in the Microwave library: we start using it to put and get microwave blobs. In reality this diff is a no-op as it still constructs the MutableRepoBlobstre from the same storage config as the non-mutable one

Reviewed By: YousefSalama

Differential Revision: D79437128

fbshipit-source-id: ffc49a98f66ea5f679a8fa4e83632f5b7dd010da