Readdir: refactor assert_attributes_result
Summary: Later in the stack, I only want to test the behavior of getAttributesFromFilesV2 (not the legacy v1 version). The wrapper expects that you always want to test v1 and v2 together. This diff refactors the wrapper so that I can use the standalone v2 wrapper for my new test.

Reviewed By: muirdm

Differential Revision: D79106405

fbshipit-source-id: 70c3f2a7f7d2f1d63cb01cdfc6cb14f413d60ae6