EntryAttributes: add additional attribute sets
Summary:
# This Diff

Add some attribute sets so that we can avoid enumerating many attributes in getEntryAttributes codepaths. This also reduces the likelihood of someone modifying (i.e. adding attributes to) one location and fogetting another location.

# Context

Currently, some clients need a combination of attributes provided by getAttributesFromFilesV2 and getFileInformation. They also need to issue readdir calls so that they can enumerate children of a given directory. This means their workflows boil down to this:

1) Issue a readdir() Thrift request
2) Parse the result as a list of files
3) Pass the entire list of files into getAttributesFromFilesV2() and issue a request for a subset of attributes. This endpoint is batching, so many file attributes can be fetched at once.
4) Then they need to separately iterate through the list of files provided by getFileInformation (since it isn't batching) and request separate attribute information for each file.
5) Once all this is done, they combine the results into a single result and process attributes for each file.

This is obviously not ideal.

Instead of forcing them to do that, we can allow users to fetch mode and mtime attributes via getAttributesFromFilesV2() so that steps 3/4/5 can all happen in a single Thrift request.

There are no technical reasons to not provide this -- just no one has gotten around to doing it yet.

Reviewed By: muirdm

Differential Revision: D79106389

fbshipit-source-id: f7739222de08fd494e081fff3c5423fe96b4dc25