return AttributeUnavailable in some cases
Summary:
# This diff

Changes some codepaths within EdenFS to return AttributeUnavailable instead of GenericError. This should help us better categorize getAttributesFromFilesV2 errors in the future.

# Context

I'd like to better categorize some errors that we're returning via Thrift. One obvious improvement we can make is differentiate between "GENERIC_ERROR" and missing attributes.

Some attributes such as Blake3, Object IDs, digest hashes, and digest sizes are expected to be unavailable in some cases. The latter 3 are always unavailable when a directory is materialized. Therefore, instead of returning GENERIC_ERROR, we can return "ATTRIBUTE_UNAVAILABLE" which indicates that the requested attribute did not exist for the given file.

ATTRIBUTE_UNAVAILABLE differs from POSIX_ERROR (ENOENT) because the latter may indicate that the tree/file itself was unavailable. In this case, it would be impossible to differentiate between the tree/file being missing and the attribute itself being missing.

ATTRIBUTE_UNAVAILABLE also differs from other unexpected errors, as other errors indicate that the attribute was expected to be available, but it was unable to be fetched due to some other issue.

Reviewed By: muirdm

Differential Revision: D79106401

fbshipit-source-id: 2cfe4d5e6aea7321d67d6a158c9332501a45677f