Readdir: add getAttributesFromFilesv2 test for materialized files/dirs
Summary:
This diff adds a new test that exercises the materialized file/dir codepath for fetching digest hashes and other attributes via getAttributesFromFilesV2. This test is important since I'm adding a new error type that we expect to be returned when a directory is materialized and digest hashes/sizes are requested via the endpoint.

In a later diff, I'll assert that the endpoint returns a AttributeUnavailable result

Reviewed By: muirdm

Differential Revision: D79106392

fbshipit-source-id: 3e45f9784a94a5480006522b72bc356b927d3088