Wire up Curl proxy support to EdenAPI
Summary:
## This stack

My ultimate goal is for a single Eden daemon to be able to support both prod and AWS repos. For that, Eden needs to be proxy-aware.

Modern Sync has a lot in common, and it's easier, so let's start there.

## This diff

EdenAPI uses Curl internally, so the first step is to set up the right setters for http proxy configuration.

Reviewed By: muirdm

Differential Revision: D79547970

fbshipit-source-id: 2444e3c2056782845921e79b1ad3fa2d1ede1862