Build a struct for EdenAPI configuration
Summary:
## This stack

My ultimate goal is for a single Eden daemon to be able to support both prod and AWS repos. For that, Eden needs to be proxy-aware.

Modern Sync has a lot in common, and it's easier, so let's start there.

## This diff

We are going to start passing more configs for EdenAPI, that's too many individual arguments.

Reviewed By: lmvasquezg

Differential Revision: D79547064

fbshipit-source-id: c20fe0cf4cdeb1764aa2ea7e7a5de8d70a8e51c7