Address dropping of exit signal sender in the sharding framework
Summary:
For background jobs, we did not need to signal termination because there was no concept of requests or a running server. Which is why we used the below pattern to stay consistent with the API:
```
    let (_, receiver) = tokio::sync::oneshot::channel::<bool>();
    executor.block_and_execute(&logger, receiver).await
```

Unfortunately, this meant that the sender gets dropped immediately after `block_and_execute`. To address this, the code has been modified to:
```
            let (sender, receiver) = tokio::sync::oneshot::channel::<bool>();
            executor.block_and_execute(&logger, receiver).await?;
            drop(sender);
            Ok(())
```

The reason we didn't detect this problem in the test plan because the test plan included canaries for services and not background jobs

Reviewed By: YousefSalama

Differential Revision: D79641513

fbshipit-source-id: d3c5256d7f4056f12f834cfccc772d71d7c04e3f