upload snippets of of rotated logs
Summary:
# This diff

Adds some logic to detect, unzip, and upload snippets of rotated/compressed logs that are created by logrotate/newsyslog tools.

# Context

We now have log rotation for EdenFS logs on macOS and Linux. In some cases, it might be worth seeing a snippet of the rotated logs to determine if they're worth collecting for futher examination.

Reviewed By: genevievehelsel

Differential Revision: D79418417

fbshipit-source-id: d3047cb2e6597d78fa3ccaa760c9394a44f6da5c