Smart actions "Powered by Devmate" icon
Summary:
Sort of experimenting / polling for feedback on this one. Typically we see the "sparkle" icon being used to indicate a "Devmate- or Metamate-powered" action, and alexchantastic suggested that we do so here as well, which might let us avoid having the "with AI" suffix on each of these actions.

It looks a bit weird having two icons side-by-side (we could alternatively just do the "sparkle" icon), so I instead put the icon after the text, and added a tooltip.

Do we have any other patterns in ISL where something like this is being done?

Reviewed By: stbfb

Differential Revision: D79202544

fbshipit-source-id: 510629955e8be7f58e96bbbf407f00bc88ab9aa2