Only render one smart actions menu at a time
Summary: The `Tooltip` component has a `group` property that will dismiss other tooltips of the same group when a new one is rendered. This is mainly used right now for preventing clutter from the different, static buttons in the top bar, but you can technically also use it for the identical, dynamic buttons that show smart actions.

Reviewed By: stbfb

Differential Revision: D79474305

fbshipit-source-id: 5b87e07f40aa0eca6460abf61c34164ecb3e5400