Disable "resolve comments" smart action if not rebased on selected commit
Summary:
The "resolve comments" action shouldn't work if you're not on the commit that the comments apply to, or else Devmate will attempt to apply the suggestions on top of changes that don't exist!

We should disable the button if we're not on the correct commit. Additionally, we can show a tooltip on the button to indicate why it's disabled.

Differential Revision: D79268579

fbshipit-source-id: ee1627cc9a1c5baabbcbb93481fa8be034045329