derivation_queue: add retry_count to DagItemInfo and increment it whenever a derivation attempt fails
Summary:
Add a `retry_count` field to `DagItemInfo` and increment it whenever a derivation attempt fails i.e. when a derivation worker calls the `nack` method on the derivation queue.

Tracking the number of retries will allow us to automatically evict items from the derivation queue if they are repeatedly failing to derive.

Reviewed By: RajivTS, lmvasquezg

Differential Revision: D79730958

fbshipit-source-id: 42d3af4c18bea8edd7d0a2769d22cc2ce924e90e