dirsync: parse `[dirsync-scripts] name = script_path`
Summary:
We are extending dirsync feature to support custom filtering. To define custom
logic, we need to add a way for the user to define a (Python) script. This diff
makes it possible to use `[dirsync-scripts]` config section
`<name> = <script_path>` for the purpose.

For security reasons (`.hgdirsync` cannot be trusted), the `[dirsync-scripts]`
in `.hgdirsync` is by default ignored, unless the repo config has
`dirsync.allow-scripts = true`.

Reviewed By: muirdm

Differential Revision: D79746229

fbshipit-source-id: ca534660e4d9574d1ac9265b41803d103a9e2b05