dirsync: split _mirrorpath to 2 functions
Summary:
In the next change, I'm adding the ability for a customized script to
decide the mirror destination path.

I'd like to avoid the ability of "escaping" the mirror destination directory.
So let's split the `mirrorpath` function so the new customized `mirrorpath` can
only decide a relative path inside the destination directory.

Reviewed By: muirdm

Differential Revision: D79746228

fbshipit-source-id: 1db653abf0f8a7cc31745405cf73f9ed6eb2dd68