Allow "resolve all comments" action to run on non-inline comments
Summary:
For the inline comments webview, each comment is known to have an associated file name and line number, so that context is always available and can be safely passed.

For the "resolve all comments" smart action, comments aren't guaranteed to be inline, and therefore may not have an associated file name and line number.

We need to adjust for this in three ways:
1. Ensure we're not passing `undefined` as a context object in the prompt
2. Update the prompt text to ensure that the model is aware that comments can be non-inline.
3. Actually pass in relevant context (in case, grab the changed files from the commit itself)

This last step is necessary, especially when all comments are non-inline, since Devmate won't have any idea where to search to make changes. For performance reasons, I'm only passing a sample of file paths, and hoping that Devmate can use this as a kernel for further searching if necessary.

Differential Revision: D79846071

fbshipit-source-id: 43cab77b9192e6a71dd410843d4225972d0e6dd6