Handle the case where tailer exits on error
Summary:
The current code assumed that ` self.tailer.run().await`  would never actually, complete. Instead we should handle the case where it completes, and run it again.

An example of this scenario can be seen in https://fb.workplace.com/groups/1708850869939124/permalink/2034976917326516/, where `run()` finished with an `Error` and the tailer stopped running

Differential Revision: D79724553

fbshipit-source-id: b8e204cba85f3c9146b1134c5e85a741cefa96bb