iobuf: drop dependency to `minibytes`
Summary:
See previous change for context. This drops the motivation to move `minibytes`
to `folly`, surfaced from sapling discord channel:

> quark — 1:16 PM
> What's the motivation behind moving minibytes? I actually prefer it being part of sapling/ for clearer ownership.
>
> benrogers — 2:02 PM
> https://github.com/search?q=repo%3Afacebook%2Ffolly%20minibytes&type=code
> (iobuf is already in folly and depends on minibytes, so moving minibytes there avoids a circular dependency)

Switching the dependency (minibytes optionally depend on iobuf) is not easy,
because iobuf does not yet have a Cargo.toml, or if it does, it is cross-repo, and
if minibytes uses a git url to refer to iobuf, the git url will prevent us from
publishing minibytes to creates.io.

Reviewed By: muirdm

Differential Revision: D79678310

fbshipit-source-id: f87765a280d6fcef4528eb3555cb1027b8193628