submodules: replace submodule array with submodule map
Summary:
Previously, only submodules of the current repo can be displayed in the repo selector.

This diff replaces the original submodule array with a map . It will help preserve submodule contexts in nested submodules, which unblocks multi-level dropdowns.

P.S. Realized trie might be a good data structure here for maps keyed by consecutive paths, but it might not worth the complexity as submodules are not accessed very intensively right now.

Reviewed By: quark-zju

Differential Revision: D79389047

fbshipit-source-id: 3ef17ecf7faaf3bd28c8016b327b7bb402897cb1