Add new entry in ISL Smart Actions for test generation
Summary:
Add a new action in SmartActions to generate tests. This is the initial integration diff. To improve the determination of changes, the 2 options are:
1. Pass additional context about the diff and changes in the initial prompt.
2. Build MCP tools to determine the current changes.

Rollout: This is controlled by the GK `isl_devmate_generate_tests_for_modified_code` which is tied to `local_test_generation_in_vs_code`. The whole `smart actions` feature is gated behind `isl_smart_actions_menu`.

Reviewed By: perehonchuk, nsblake

Differential Revision: D79885903

fbshipit-source-id: 2cf87ba749d9b8cca8282a443f252e4d6378955b