Commit metadata scraper binary 2/n - fill impl
Summary:
In this stack, I'll gradually convert the existing Hack script DumpCommitsMetadata to a binary so I have better control on where the derivation for draft commit goes.
 ---
This diff filled up the computation logic for the remaining fields. The goal is to keep it consistent with the existing Hack solution.

Reviewed By: clara-9

Differential Revision: D79297893

fbshipit-source-id: 26b4bc6b4dc26341ae8da746f4dd177e494dcc0f