CwdSelector: trim leading seperator in relativePath()
Summary:
In `relativePath()`, we calls `path.replace` to remove matching prefixes.

However, this sometimes leaves a leading separator when the root path doesn't end with one. For example, `relativePath('/root', '/root/dir')` returns `/dir`, which is not really a desirable form of a relative path. And this is an issue for matching submodules since normally they don't have a leading dash.

This diff enables `relativePath()` to automatically trim leading dash before returning results.

Reviewed By: quark-zju

Differential Revision: D79775705

fbshipit-source-id: 907f384d4428fdd62ca4079c46e9f343d0104e36