export: pass entire FileInfo into export_file()
Summary:
I want to add additional ways to filter out files when exporting, so first tweak things to pass along the entire FileInfo object to export_file() (which currently does the filtering).

Note that, as I've recently discovered, the FileInfo is not fully populated from the tree_list() SCS endpoint.

Reviewed By: chiara125

Differential Revision: D79288477

fbshipit-source-id: e05e1294a05f77023e21c7213a90a069b9cc3044