revisionstore: support uncompressed content in IndexedLogHgIdDataStore
Summary:
In preparation for (sometimes) storing uncompressed tree and file data in indexedlog, add read/write support for uncompressed data.

When storing uncompressed data, we add a new 'u' metadata flag in the serialized metadata. When reading, if there is a 'u', we don't decompress.

I wrapped the existing Metadata into a new struct "InternalMetadata", containing the new flag. This way we don't have to expose the "is uncompressed" flag to external users of the Metadata type.

In this commit we never write uncompressed data since that would be incompatible with old code. Once this commit has been deployed for a while we can start writing uncompressed data.

Reviewed By: quark-zju

Differential Revision: D79932147

fbshipit-source-id: 16bdb912816bf66e055b4082aa08fea350c74d68