Mononoke: Add with_cmd_args to allow creating a MononokeApp with args that don't come from the cmdline
Summary:
I need this to have tests that validate command line parsing including the args that MononokeApp appends on its own.

I can't change the env variable from the test, so I need a way to pass my own args to the app.

I think this is clean enough to add that without uglyfying Mononoke.

Reviewed By: markbt

Differential Revision: D80118394

fbshipit-source-id: c3c00c2b2547549162b6b7606b6f50b5706d0e4e