derived_data: implement blocking of derivation in config
Summary:
We would like to have the ability to blocklist specific commits from attempting derivation.  This can be used as a mitigation during SEVs or to prevent excess work when we attempt to derive known bad commits.

This diff implements blocking of derivation in config.  We check for blocked derivation in two places:
* Before requesting remote derivation.  This avoids the request (and the load on the derived data service) in the case that the requested changeset is blocked.  It can only check the requested changeset, not any of the dependencies as they haven't been found yet.
* When performing derivation for each commit.  This is necessary to catch derivation triggered by a descendant of a blocked commit.  For remote derivation we will rely on the recently added "attempts" counter to ensure that we don't continually attempt derivation on derivation workers.

Reviewed By: YousefSalama

Differential Revision: D80045900

fbshipit-source-id: a4ce2b52bfb4d845da3ef4c494f2be0428ac6861