--upload argument for `eden debug log`
Summary:
`eden debug log --upload` is not working.
```
[kavehahmadi@devvm8414.ash0 ~/fbsource/fbcode (533435a9ef)]$ eden debug log --upload
P1904358854: https://www.internalfb.com/intern/paste/P1904358854/
```
The output in the paste is an error.

Recently we changed the `print_log_file` function. Updating the argument list for this function. Therefore, the `open_fn` will get the default value which is a callable function.

Reviewed By: genevievehelsel

Differential Revision: D80273163

fbshipit-source-id: 3425072119402d3569c654eaa74a44b85b4cb560