Implement memory based load balancing for Git Locator
Summary:
We were hoping to rely on the memory based load balancing support provided by Service Router, but the responses on this post of mine made it clear that it won't work.

So we are left with only one option which is to rollout our own memory based load balancing for Mononoke Git. This diff creates a basic memory based load balancer and we can definitely make iterative improvements on it, but this is a good enough first version.

The change is guarded by a JK that has been already created and set to false

{F1981222175}

NOTE: There is a perf concern here because fetching the mem utilisation of relevant hosts for every request can be expensive (e.g. 0.5 secs). For Git requests, this is a worthy trade off because all Git requests are in the order of minutes or hours. However, for SLAPI in the future this might not make sense in which case we can perform endpoint specific load balancing

Reviewed By: YousefSalama

Differential Revision: D80275419

fbshipit-source-id: 27c84944b03a7c80b619ef033f69ac4d8b997fa2