advertise bundle uri based on repo config rather than the JK
Summary:
This should be a noop moving the functionality from the GK to repo config in configerator.

If someone passes an explicit flag to use or not use bundles that always takes precedence.

Otherwise we should use the config from configerator (already synced with the JK).

We use the same config to configure bundle generation for a repo, so this means that we will be serving temporarily empty bundle lists for a repo via bundle-uri functionality, until the bundle-generator generated the bundle. This is fine.

Reviewed By: andreacampi

Differential Revision: D80260858

fbshipit-source-id: 1039f38250510af8732a0c8ef26b4eb04e0e0c19