Support both lowercase and uppercase proxy env variables
Summary:
D66759176 added support for `HTTP_PROXY`, which is less widely used than `http_proxy`. Importantly, the latter is what `curl` and `libcurl` use.

D79547970 added support for `NO_PROXY`, which is also uncommon.

This diff adds support for the lowercase variants. It also maintains support for the uppercase variant, because they are still widespread even if minority.

Reviewed By: markbt

Differential Revision: D80438780

fbshipit-source-id: 202112b7bdbbb263796ad26a1229db71a5a3f5c5