add a new column to silent exit in edenfs_events
Summary:
## This diff
Add a new column to `silent_daemon_exit` event in edenfs_events table
## Context
By auditing the macOS users silent exit, I found that a number of users experienced silent exit after a force shutdown or power outage due to dead battery. These events are not related to eden and should not consider as SAD events.
To exclude these events from %SAD metric, we need more data in our scuba logs. We already have the last edenfs heartbeat time in our logs. It shows what was the last time that previous eden was alive. If we know the last time that mac host rebooted then we can compare the two timestamp, and find if the silent exit was due to a battery outage and system reboot or no system reboot happen since the last time that eden existed silently, then the silent exit could be related to other cause such as OOM kill.

Differential Revision: D80379424

fbshipit-source-id: 63bd5346490698ed3bca2629323f9800022536f1